refac: Internal google API Service 분리
