+++ src/main/java/com/jjbacsa/jjbacsabackend/google/serviceImpl/InternalGoogleServiceImpl.java
@@ -1,22 +1,13 @@
 package com.jjbacsa.jjbacsabackend.google.serviceImpl;
 
 import com.jjbacsa.jjbacsabackend.etc.enums.ErrorMessage;
-import com.jjbacsa.jjbacsabackend.etc.exception.ApiException;
-import com.jjbacsa.jjbacsabackend.etc.exception.BaseException;
 import com.jjbacsa.jjbacsabackend.etc.exception.RequestInputException;
-import com.jjbacsa.jjbacsabackend.google.dto.response.ShopScrapResponse;
 import com.jjbacsa.jjbacsabackend.google.entity.GoogleShopEntity;
 import com.jjbacsa.jjbacsabackend.google.repository.GoogleShopRepository;
-import com.jjbacsa.jjbacsabackend.google.dto.response.ShopResponse;
-import com.jjbacsa.jjbacsabackend.google.service.GoogleShopService;
 import com.jjbacsa.jjbacsabackend.google.service.InternalGoogleService;
-import com.jjbacsa.jjbacsabackend.scrap.entity.ScrapEntity;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -24,20 +15,6 @@ import java.util.stream.Collectors;
 public class InternalGoogleServiceImpl implements InternalGoogleService {
 
     private final GoogleShopRepository googleShopRepository;
-    private final GoogleShopService googleService;
-
-    @Override
-    public GoogleShopEntity getGoogleShopByPlaceId(String placeId) {
-
-        return googleShopRepository.findByPlaceId(placeId)
-                .orElseGet(() -> {
-                    try {
-                        return saveGoogleShop(placeId);
-                    } catch (Exception e) {
-                        throw new ApiException(ErrorMessage.NOT_FOUND_EXCEPTION);
-                    }
-                });
-    }
 
     @Override
     @Transactional(readOnly = true)
@@ -45,12 +22,6 @@ public class InternalGoogleServiceImpl implements InternalGoogleService {
 
         return googleShopRepository.findById(shopId)
                 .orElseThrow(() -> new RequestInputException(ErrorMessage.SHOP_NOT_EXISTS_EXCEPTION));
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public ShopResponse getShopDetails(String placeId) throws Exception {
-        return googleService.getShopDetails(placeId, true);
     }
 
     @Override
@@ -74,18 +45,4 @@ public class InternalGoogleServiceImpl implements InternalGoogleService {
         shop.getShopCount().setRatingCount(googleShopRepository.getRatingCount(shopId) - 1);
     }
 
-    private GoogleShopEntity saveGoogleShop(String placeId) throws Exception {
-        ShopResponse shopResponse = googleService.getShopDetails(placeId, true);
-        GoogleShopEntity googleShopEntity = GoogleShopEntity.builder()
-                .placeId(shopResponse.getPlaceId())
-                .build();
-
-        return googleShopRepository.save(googleShopEntity);
-    }
-
-    @Override
-    public ShopScrapResponse formattedToShopResponse(ScrapEntity scrap) throws Exception {
-        ShopScrapResponse shopScrapResponse = this.googleService.getShopScrap(scrap.getShop().getPlaceId(), scrap.getId());
-        return shopScrapResponse;
-    }
 }
